requirements.txt
openai
httpx
python-dotenv
requests

